refactor: aplica sugestões de code review do Vini no PR #22
- renomeia src/config.py para src/constants.py
- centraliza constantes (SMTP_HOST, SMTP_PORT, TIMEOUT_SEGUNDOS)
- extrai JS inline para src/js/list_values.js
- atualiza documentacao.md com nova estrutura
