--- src/config.py
@@ -1,9 +1,0 @@
-"""
-Configurações globais da aplicação.
-"""
-
-# URL fixa do Google Form onde as alterações serão registradas
-FORM_URL = "https://forms.gle/9qC2PVjEdfA3QQ7E9"
-
-# Intervalo de monitoramento em segundos
-INTERVALO_SEGUNDOS = 15
