+++ src/constants.py
@@ -1,0 +1,16 @@
+"""
+Constantes globais da aplicação.
+"""
+
+# URL fixa do Google Form onde as alterações serão registradas
+FORM_URL = "https://forms.gle/9qC2PVjEdfA3QQ7E9"
+
+# Intervalo de monitoramento em segundos
+INTERVALO_SEGUNDOS = 15
+
+# Timeout em segundos para esperas do Selenium (ex: carregamento do form)
+TIMEOUT_SEGUNDOS = 10
+
+# Configurações do servidor SMTP (Gmail)
+SMTP_HOST = "smtp.gmail.com"
+SMTP_PORT = 587
